Move highlighting code to separate file
