+++ tools/shell/linenoise/include/highlighting.hpp
@@ -11,6 +11,7 @@
 #include "duckdb/common/common.hpp"
 
 namespace duckdb {
+struct searchMatch;
 
 enum class tokenType : uint8_t {
 	TOKEN_IDENTIFIER,
@@ -29,8 +30,19 @@ struct highlightToken {
 	bool search_match = false;
 };
 
-class LinenoiseHighlighting {
+class Highlighting {
+public:
+	static void Enable();
+	static void Disable();
+	static bool IsEnabled();
+	static const char *GetColorOption(const char *option);
+	static void SetKeyword(const char *color);
+	static void SetConstant(const char *color);
 
-};
+	static vector<highlightToken> Tokenize(char *buf, size_t len, searchMatch *match = nullptr);
+	static string HighlightText(char *buf, size_t len, size_t start_pos, size_t end_pos,
+							  const vector<highlightToken> &tokens);
+
+	};
 
 } // namespace duckdb
